remove invalid errors on console, remove logger, move app.js to containers folder, it connects to redux
