+++ src/index.js
@@ -2,15 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import logger from 'redux-logger';
 import { Provider } from 'react-redux';
 import thunkMiddleware from 'redux-thunk';
 import indexReducer from './reducers/index';
-import App from './components/App';
+import App from './containers/App';
 import './index.css';
 
 const store = createStore(indexReducer, composeWithDevTools(
-  applyMiddleware(thunkMiddleware, logger),
+  applyMiddleware(thunkMiddleware),
 ));
 
 ReactDOM.render(
